+++ template/src/js/app/main/index.js
@@ -1,10 +1,8 @@
-import '../../../css/common';
-import 'common/const';
-import 'common/interface';
-$.ajax({
-    url: '/api/patch/gettestchannelId',
-    type: 'get',
-    success: function(res) {
-        console.log(res);
+import Vue from 'vue';
+
+new Vue({
+    el: '#app',
+    data: {
+        text: 'hello world'
     }
 });